requirements.txt
blinker==1.9.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
Flask==3.0.3
imbalanced-learn==0.14.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
numpy==2.4.3
oracledb==3.4.2
pandas==3.0.1
pycparser==3.0
python-dateutil==2.9.0.post0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sklearn-compat==0.1.5
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
Werkzeug==3.1.6
xgboost==3.2.0
